perf: lazy-load three.js DottedSurface background via next/dynamic
