+++ components/ui/DottedSurfaceCanvas.tsx
@@ -1,0 +1,148 @@
+'use client';
+import { cn } from '@/lib/utils';
+import React, { useEffect, useRef } from 'react';
+import * as THREE from 'three';
+
+export type DottedSurfaceCanvasProps = Omit<React.ComponentProps<'div'>, 'ref'>;
+
+export function DottedSurfaceCanvas({ className, ...props }: DottedSurfaceCanvasProps) {
+
+    const containerRef = useRef<HTMLDivElement>(null);
+    const sceneRef = useRef<{
+        scene: THREE.Scene;
+        camera: THREE.PerspectiveCamera;
+        renderer: THREE.WebGLRenderer;
+        particles: THREE.Points[];
+        animationId: number;
+        count: number;
+    } | null>(null);
+
+    useEffect(() => {
+        if (!containerRef.current) return;
+
+        const SEPARATION = 150;
+        const AMOUNTX = 40;
+        const AMOUNTY = 60;
+
+        const scene = new THREE.Scene();
+        scene.fog = new THREE.Fog(0xffffff, 2000, 10000);
+
+        const camera = new THREE.PerspectiveCamera(
+            60,
+            window.innerWidth / window.innerHeight,
+            1,
+            10000,
+        );
+        camera.position.set(0, 355, 1220);
+
+        const renderer = new THREE.WebGLRenderer({
+            alpha: true,
+            antialias: true,
+        });
+        renderer.setPixelRatio(window.devicePixelRatio);
+        renderer.setSize(window.innerWidth, window.innerHeight);
+        renderer.setClearColor(scene.fog.color, 0);
+
+        containerRef.current.appendChild(renderer.domElement);
+
+        const positions: number[] = [];
+        const colors: number[] = [];
+        const geometry = new THREE.BufferGeometry();
+
+        for (let ix = 0; ix < AMOUNTX; ix++) {
+            for (let iy = 0; iy < AMOUNTY; iy++) {
+                positions.push(
+                    ix * SEPARATION - (AMOUNTX * SEPARATION) / 2,
+                    0,
+                    iy * SEPARATION - (AMOUNTY * SEPARATION) / 2,
+                );
+                colors.push(0.78, 0.78, 0.78);
+            }
+        }
+
+        geometry.setAttribute('position', new THREE.Float32BufferAttribute(positions, 3));
+        geometry.setAttribute('color', new THREE.Float32BufferAttribute(colors, 3));
+
+        const material = new THREE.PointsMaterial({
+            size: 8,
+            vertexColors: true,
+            transparent: true,
+            opacity: 0.45,
+            sizeAttenuation: true,
+        });
+
+        const points = new THREE.Points(geometry, material);
+        scene.add(points);
+
+        let count = 0;
+        let animationId = 0;
+
+        const animate = () => {
+            animationId = requestAnimationFrame(animate);
+
+            const positionAttribute = geometry.attributes.position;
+            const pos = positionAttribute.array as Float32Array;
+
+            let i = 0;
+            for (let ix = 0; ix < AMOUNTX; ix++) {
+                for (let iy = 0; iy < AMOUNTY; iy++) {
+                    pos[i * 3 + 1] =
+                        Math.sin((ix + count) * 0.3) * 50 +
+                        Math.sin((iy + count) * 0.5) * 50;
+                    i++;
+                }
+            }
+
+            positionAttribute.needsUpdate = true;
+            renderer.render(scene, camera);
+            count += 0.1;
+        };
+
+        const handleResize = () => {
+            camera.aspect = window.innerWidth / window.innerHeight;
+            camera.updateProjectionMatrix();
+            renderer.setSize(window.innerWidth, window.innerHeight);
+        };
+
+        window.addEventListener('resize', handleResize);
+        const reducedMotion = window.matchMedia('(prefers-reduced-motion: reduce)').matches;
+        if (reducedMotion) {
+            renderer.render(scene, camera);
+        } else {
+            animate();
+        }
+
+        sceneRef.current = { scene, camera, renderer, particles: [points], animationId, count };
+
+        const container = containerRef.current;
+
+        return () => {
+            window.removeEventListener('resize', handleResize);
+            if (sceneRef.current) {
+                cancelAnimationFrame(sceneRef.current.animationId);
+                sceneRef.current.scene.traverse((object) => {
+                    if (object instanceof THREE.Points) {
+                        object.geometry.dispose();
+                        if (Array.isArray(object.material)) {
+                            object.material.forEach((m) => m.dispose());
+                        } else {
+                            object.material.dispose();
+                        }
+                    }
+                });
+                sceneRef.current.renderer.dispose();
+                if (container && sceneRef.current.renderer.domElement) {
+                    container.removeChild(sceneRef.current.renderer.domElement);
+                }
+            }
+        };
+    }, []);
+
+    return (
+        <div
+            ref={containerRef}
+            className={cn('pointer-events-none fixed inset-0', className)}
+            {...props}
+        />
+    );
+}
